Almost Done

+++ Controllers/ProjectsController.cs
@@ -19,33 +19,35 @@ namespace Portofolio.Controllers
 {
     public class ProjectsController : Controller
     {
-        private readonly IRepository<Project> _projectsRepository;
+        private readonly ProjectsRepository _projectsRepository;
 
         private readonly UserManager<User> _userManager;
-        private readonly IRepository<UsersInProject> _uipRepository;
-        private readonly IRepository<ProjectType> _projectTypeRepository;
+        private readonly UIPRepository _uipRepository;
+        private readonly ProjectTypeRepository _projectTypeRepository;
 
-        private readonly IRepository<LinkType> _linkTypesRepository;
+        private readonly LinkTypesRepository _linkTypesRepository;
 
-        private readonly IRepository<UserRoleInProject> _uripRepository;
+        private readonly UserRoleInProjectRepository _uripRepository;
 
-        private readonly IRepository<ProjectLink> _projectLinksRepository;
+        private readonly ProjectLinksRepository _projectLinksRepository;
 
-        private readonly IRepository<ProjectImage> _projectImagesRepository;
+        private readonly ProjectImagesRepository _projectImagesRepository;
 
         private readonly IImageServices _imageServices;
 
-        private readonly IRepository<ProjectFeedback> _feedbackRepository;
+        private readonly ProjectFeedbacksRepository _feedbackRepository;
 
         private readonly IPaginator<Project> _projectPaginator;
 
-        private readonly IRepository<ProjectLog> _projectLogRepository;
+        private readonly ProjectLogsRepository _projectLogRepository;
 
-        private readonly IRepository<Contact> _contactsRepository;
+        private readonly ContactRepository _contactsRepository;
 
         private readonly IDisplayOutput _outputDisplayer;
 
-        public ProjectsController(IDisplayOutput outputDisplayer, BaseRepository<Contact> contactsRepository, BaseRepository<ProjectLog> projectLogRepository, BasePaginator<Project> projectPaginator, BaseRepository<ProjectFeedback> feedbackRepository, BaseRepository<ProjectImage> projectImagesRepository, BaseRepository<ProjectLink> projectLinksRepository, BaseRepository<UsersInProject> uipRepository, IImageServices imageServices, BaseRepository<UserRoleInProject> uripRepository, BaseRepository<LinkType> linkTypesRepository, BaseRepository<Project> projectsRepository, UserManager<User> userManager, BaseRepository<ProjectType> projectTypeRepository)
+        private readonly ServicesRepository _servicesRepository;
+
+        public ProjectsController(ServicesRepository servicesRepository, IDisplayOutput outputDisplayer, ContactRepository contactsRepository, ProjectLogsRepository projectLogRepository, BasePaginator<Project> projectPaginator, ProjectFeedbacksRepository feedbackRepository, ProjectImagesRepository projectImagesRepository, ProjectLinksRepository projectLinksRepository, UIPRepository uipRepository, IImageServices imageServices, UserRoleInProjectRepository uripRepository, LinkTypesRepository linkTypesRepository, ProjectsRepository projectsRepository, UserManager<User> userManager, ProjectTypeRepository projectTypeRepository)
         {
             _userManager = userManager;
             _projectsRepository = projectsRepository;
@@ -61,25 +63,37 @@ namespace Portofolio.Controllers
             _projectLogRepository = projectLogRepository;
             _contactsRepository = contactsRepository;
             _outputDisplayer = outputDisplayer;
+            _servicesRepository = servicesRepository;
         }
 
         public async Task<IActionResult> Index(int page = 1)
         {
+            var latestProjects = await _projectsRepository.GetLatestProjects(4);
+            var latestServices = await _servicesRepository.GetLatestServices(4);
             var projects = await _projectsRepository.GetAll();
             var paginationVM = _projectPaginator.Paginate(projects, page);
-            return View(paginationVM);
+            return View(new ProjectsViewModel
+            {
+                LatestProjects = latestProjects,
+                LatestServices = latestServices,
+                PaginationModel = paginationVM
+            });
         }
 
         public async Task<IActionResult> Details(int id)
         {
-            var project = await _projectsRepository.GetById(id);
-            if(project == default(Project))
+            var latestProjects = await _projectsRepository.GetLatestProjects(4);
+            var latestServices = await _servicesRepository.GetLatestServices(4);
+            var project = await _projectsRepository.GetByIdWithInclude(id);
+            if (project == default(Project))
             {
-                return RedirectToAction(nameof(ErrorsController.Error404), "Error");
+                return NotFound();
             }
             return View(new ProjectDetailsViewModel
             {
-                Project = await _projectsRepository.GetById(id)
+                Project = await _projectsRepository.GetByIdWithInclude(id),
+                LatestProjects = latestProjects,
+                LatestServices = latestServices
             });
         }
 
@@ -95,7 +109,7 @@ namespace Portofolio.Controllers
                 ProjectId = model.ProjectId,
                 Feedback = model.Feedback
             });
-            _outputDisplayer.DisplayOutput(ViewData, true, "Feedback submitted successfully");
+            _outputDisplayer.DisplayOutput(TempData, true, "Feedback submitted successfully");
             return RedirectToAction(nameof(Details), new { id = model.ProjectId });
         }
 
@@ -107,7 +121,7 @@ namespace Portofolio.Controllers
             var projectTypes = await _projectTypeRepository.GetAll();
             var linkTypes = await _linkTypesRepository.GetAll();
             var urips = await _uripRepository.GetAll();
-            var pendingContactsCount = (await _contactsRepository.GetAll()).Where((contact) => contact.Status.Status == ContactStatuses.Pending.ToString()).Count();
+            var pendingContactsCount = await _contactsRepository.PendingContactsCount();
             return View(new CreateProjectViewModel
             {
                 PendingContactsCount = pendingContactsCount,
@@ -129,15 +143,17 @@ namespace Portofolio.Controllers
                 ModelState.AssignViewDataWithErrors(ViewData);
                 return RedirectToAction(nameof(Create));
             }
-            var teamLeaderRole = await _uripRepository.FindByCondition((urip) => urip.Role == TeamLeaderRole);
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+            var teamLeaderRole = await _uripRepository.GetTeamLeaderRole();
             var teamLeader = projectViewModel.UsersNRolesIds.Where((usersNRoleId) => Convert.ToInt32(usersNRoleId.Split('-')[0]) == teamLeaderRole.Id).FirstOrDefault();
-            if(teamLeader == default(string))
+            if (teamLeader == default(string))
             {
-                _outputDisplayer.DisplayOutput(ViewData, false, "Please select a team leader for the project");
+                _outputDisplayer.DisplayOutput(TempData, false, "Please select a team leader for the project");
                 return RedirectToAction(nameof(Create));
             }
             var project = await _projectsRepository.Create(new Project
             {
+                CreatorId = user.Id,
                 Title = projectViewModel.Title,
                 Description = projectViewModel.Description,
                 TypeId = projectViewModel.ProjectTypeId,
@@ -148,16 +164,16 @@ namespace Portofolio.Controllers
             try
             {
                 _imageServices.ValidateImgExtension(projectViewModel.ProjectThumbnail);
-                project.Thumbnail = await _imageServices.UploadImgAsync(projectViewModel.ProjectThumbnail);
-                await _imageServices.ResizeImg(project.Thumbnail, ProjectImageSize);
-                await _projectsRepository.Edit(project);
+                var thumbnailPath = await _imageServices.UploadImgAsync(projectViewModel.ProjectThumbnail);
+                await _imageServices.ResizeImg(thumbnailPath, ProjectImageSize);
+                await _projectsRepository.AddThumbnail(project.Id, thumbnailPath);
             }
             catch (CustomException ex)
             {
-                _outputDisplayer.DisplayOutput(ViewData, false, ex.Message);
+                _outputDisplayer.DisplayOutput(TempData, false, ex.Message);
                 return RedirectToAction(nameof(Create));
             }
-            if (projectViewModel.ProjectImages != null)
+            if (projectViewModel.ProjectImages != default(IFormFileCollection))
             {
                 try
                 {
@@ -165,13 +181,13 @@ namespace Portofolio.Controllers
                 }
                 catch (CustomException ex)
                 {
-                    _outputDisplayer.DisplayOutput(ViewData, false, ex.Message);
+                    _outputDisplayer.DisplayOutput(TempData, false, ex.Message);
                     return RedirectToAction(nameof(Create));
                 }
             }
-            _outputDisplayer.DisplayOutput(ViewData, true, "Project created successfully");
-            var user = await _userManager.GetUserAsync(HttpContext.User);
-            await _projectLogRepository.Create(new ProjectLog{
+            _outputDisplayer.DisplayOutput(TempData, true, "Project created successfully");
+            await _projectLogRepository.Create(new ProjectLog
+            {
                 ProjectId = project.Id,
                 UserId = user.Id,
                 Title = project.Title,
@@ -189,23 +205,29 @@ namespace Portofolio.Controllers
             if (!ModelState.IsValid)
             {
                 ModelState.AssignViewDataWithErrors(ViewData);
-                return RedirectToAction(nameof(DashboardController.ProjectDetails), new {id = editViewModel.ProjectId});
+                return RedirectToAction(nameof(DashboardController.ProjectDetails), new { id = editViewModel.ProjectId });
             }
-            var edittedProject = await _projectsRepository.Edit(new Project
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+            var project = await _projectsRepository.GetByIdNoInclude(editViewModel.ProjectId);
+            if (user.Id != project.CreatorId && !await _userManager.IsInRoleAsync(user, UserRoles.Admin.ToString()))
             {
-                Id = editViewModel.ProjectId,
+                return Unauthorized();
+            }
+            var edittedProject = await _projectsRepository.Edit(editViewModel.ProjectId, new Project
+            {
                 Title = editViewModel.Title,
                 Description = editViewModel.Description,
                 ToolsUsed = editViewModel.ToolsUsed,
                 TypeId = editViewModel.ProjectTypeId
             });
+            edittedProject.UsersInProjects = await _uipRepository.GetProjectUips(edittedProject.Id);
             await _uipRepository.EditFromCollection(edittedProject.UsersInProjects, editViewModel.UsersNRolesIds, '-');
-            if (editViewModel.AddUsersNRolesIds != null && editViewModel.AddUsersNRolesIds.Count != 0)
+            if (editViewModel.AddUsersNRolesIds != default(ICollection<string>) && editViewModel.AddUsersNRolesIds.Count != 0)
             {
                 await _uipRepository.CreateFromCollection(editViewModel.ProjectId, editViewModel.AddUsersNRolesIds, '-');
             }
             await _projectLinksRepository.EditFromCollection(editViewModel.LinksIds, editViewModel.ProjectLinks);
-            if (editViewModel.ProjectThumbnail != null)
+            if (editViewModel.ProjectThumbnail != default(IFormFile))
             {
                 try
                 {
@@ -213,13 +235,12 @@ namespace Portofolio.Controllers
                     string imagePath = await _imageServices.UploadImgAsync(editViewModel.ProjectThumbnail);
                     await _imageServices.ResizeImg(imagePath, ProjectImageSize);
                     _imageServices.DeleteImg(edittedProject.Thumbnail);
-                    edittedProject.Thumbnail = imagePath;
-                    await _projectsRepository.Edit(edittedProject);
+                    await _projectsRepository.AddThumbnail(edittedProject.Id, imagePath);
                 }
                 catch (CustomException ex)
                 {
-                    _outputDisplayer.DisplayOutput(ViewData, false, ex.Message);
-                    return RedirectToAction(nameof(DashboardController.ProjectDetails), new {id = editViewModel.ProjectId});
+                    _outputDisplayer.DisplayOutput(TempData, false, ex.Message);
+                    return RedirectToAction(nameof(DashboardController.ProjectDetails), new { id = editViewModel.ProjectId });
                 }
             }
             if (editViewModel.ProjectImages != default(IFormFileCollection) && editViewModel.ProjectImages.Count != 0)
@@ -230,13 +251,13 @@ namespace Portofolio.Controllers
                 }
                 catch (CustomException ex)
                 {
-                    _outputDisplayer.DisplayOutput(ViewData, false, ex.Message);
-                    return RedirectToAction(nameof(DashboardController.ProjectDetails), new {id = editViewModel.ProjectId});
+                    _outputDisplayer.DisplayOutput(TempData, false, ex.Message);
+                    return RedirectToAction(nameof(DashboardController.ProjectDetails), new { id = editViewModel.ProjectId });
                 }
             }
-            _outputDisplayer.DisplayOutput(ViewData, true, "Project edit is successful");
-            var user = await _userManager.GetUserAsync(HttpContext.User);
-            await _projectLogRepository.Create(new ProjectLog{
+            _outputDisplayer.DisplayOutput(TempData, true, "Project edit is successful");
+            await _projectLogRepository.Create(new ProjectLog
+            {
                 ProjectId = edittedProject.Id,
                 UserId = user.Id,
                 Title = edittedProject.Title,
@@ -249,7 +270,12 @@ namespace Portofolio.Controllers
         [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
-            var project = await _projectsRepository.GetById(id);
+            var project = await _projectsRepository.GetByIdWithInclude(id);
+            var user = await _userManager.GetUserAsync(HttpContext.User);
+            if (user.Id != project.CreatorId && !await _userManager.IsInRoleAsync(user, UserRoles.Admin.ToString()))
+            {
+                return Unauthorized();
+            }
             if (project.ProjectFeedbacks != default(ICollection<ProjectFeedback>) && project.ProjectFeedbacks.Count != 0)
             {
                 await _feedbackRepository.DeleteCollection(project.ProjectFeedbacks);
@@ -261,9 +287,9 @@ namespace Portofolio.Controllers
             await _projectImagesRepository.DeleteCollection(project.ProjectImages);
             await _projectLinksRepository.DeleteCollection(project.ProjectLinks);
             await _uipRepository.DeleteCollection(project.UsersInProjects);
-            await _projectsRepository.Delete(project);
-            var user = await _userManager.GetUserAsync(HttpContext.User);
-            await _projectLogRepository.Create(new ProjectLog{
+            await _projectsRepository.Delete(project.Id);
+            await _projectLogRepository.Create(new ProjectLog
+            {
                 ProjectId = id,
                 UserId = user.Id,
                 Title = project.Title,
@@ -276,12 +302,17 @@ namespace Portofolio.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteImage(int id)
         {
-            var image = await _projectImagesRepository.GetById(id);
-            _imageServices.DeleteImg(image.ImagePath);
-            await _projectImagesRepository.Delete(image);
             var user = await _userManager.GetUserAsync(HttpContext.User);
-            var project = await _projectsRepository.GetById(image.ProjectId);
-            await _projectLogRepository.Create(new ProjectLog{
+            var image = await _projectImagesRepository.GetById(id);
+            var project = await _projectsRepository.GetByIdNoInclude(image.ProjectId);
+            if (user.Id != project.CreatorId && !await _userManager.IsInRoleAsync(user, UserRoles.Admin.ToString()))
+            {
+                return Unauthorized();
+            }
+            _imageServices.DeleteImg(image.ImagePath);
+            await _projectImagesRepository.Delete(image.Id);
+            await _projectLogRepository.Create(new ProjectLog
+            {
                 ProjectId = project.Id,
                 UserId = user.Id,
                 Title = project.Title,
@@ -294,18 +325,22 @@ namespace Portofolio.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteUip(int id)
         {
-            var uip = await _uipRepository.GetById(id);
-            await _uipRepository.Delete(uip);
-            var project = await _projectsRepository.GetById(uip.ProjectId);
+            var uip = await _uipRepository.Delete(id);
+            var project = await _projectsRepository.GetByIdNoInclude(uip.ProjectId);
             var user = await _userManager.GetUserAsync(HttpContext.User);
-            await _projectLogRepository.Create(new ProjectLog{
-                ProjectId = project.Id,
+            if (user.Id != project.CreatorId && !await _userManager.IsInRoleAsync(user, UserRoles.Admin.ToString()))
+            {
+                return Unauthorized();
+            }
+            await _projectLogRepository.Create(new ProjectLog
+            {
+                ProjectId = uip.ProjectId,
                 UserId = user.Id,
                 Title = project.Title,
                 Action = ProjectUipDeleteAction,
                 UserName = user.UserName
             });
-            return RedirectToAction(nameof(DashboardController.ProjectDetails), "Dashboard", new {id = uip.ProjectId});
+            return RedirectToAction(nameof(DashboardController.ProjectDetails), "Dashboard", new { id = uip.ProjectId });
         }
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
